+++ src/Coypu/State.cs
@@ -5,15 +5,20 @@ namespace Coypu
     ///<summary>
     /// A possible state for the current page
     ///</summary>
-    public class State
+    public class State 
     {
-        public Func<bool> Condition { get; set; }
+        private Func<bool> condition;
+
+        public State(Func<bool> condition) 
+        {
+            this.condition = condition;
+        }
 
         public bool ConditionMet { get; private set; }
 
         public bool CheckCondition()
         {
-            return ConditionMet = Condition();
+            return ConditionMet = condition();
         }
     }
 }